requirements.txt
numpy
pandas
requests
matplotlib
jupyter
pytest
ttkbootstrap
keyboard
pillow
psutil
pyyaml
pygetwindow
pywin32
pydirectinput
pyautogui
opencv-python
mss
easyocr